Add authentication system & some styling on forms

+++ src/Controller/PinsController.php
@@ -91,13 +91,16 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]+>}", name="app_pins_delete", methods={"DELETE"})
      */
-    public function delete(Pin $pin, EntityManagerInterface $em) : Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em) : Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
 
-        $this->addFlash('info', 'Pin successfully deleted');
+            $this->addFlash('info', 'Pin successfully deleted');
 
+        }
+        
 
         return $this->redirectToRoute('app_home');
     }
